refactor of modal wrapper
